requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-dotenv>=1.0.0
